Backend/sdp_backend/requirements.txt
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
bcrypt==3.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.6
django-cors-headers==3.0.2
django-filter==2.1.0
django-queryset-csv==1.0.2
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-jsonp==1.0.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
gitdb2==4.0.2
GitPython==3.0.6
idna==2.8
itypes==1.2.0
multidict==5.1.0
oauthlib==3.1.0
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.22.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
soupsieve==2.1
sqlparse==0.4.1
tldextract==2.2.0
truffleHog==2.1.11
truffleHogRegexes==0.0.7
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.25.11
yarl==1.6.3
